% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\docType{data}
\name{hdi}
\alias{hdi}
\title{hdi
Human Development Index data by UNDP for 188 country from 1990 to 2015.}
\format{The dataframe \code{hdi} has 4888 rows and 3 variables
\itemize{
\item \strong{country}: factor with 188 levels
\item \strong{year}: years between 1900 - 2015
\item \strong{human_development_index}: measured index with lower values reflecting lower development
}

The data originally comes in wide format \code{hdi_wide}}
\source{
\itemize{
\item \href{http://hdr.undp.org/en/data}{UNDP}
}
}
\usage{
hdi
}
\description{
hdi
Human Development Index data by UNDP for 188 country from 1990 to 2015.
}
\keyword{datasets}
