% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_cpi_2015.R
\docType{data}
\name{hdi_cpi_2015}
\alias{hdi_cpi_2015}
\title{hdi_cpi_2015 Data}
\format{The main dataframe \code{hdi_cpi_2015} has 121 rows and 10 variables:
\describe{
\item{country}{factor with 121 levels}
\item{country_code}{factor}
\item{region}{factor with 6 levels}
\item{continent}{factor with 5 levels}
\item{population}{country population in 2015}
\item{corruption_perception_index}{lower values reflect higher perceived corruption}
\item{cpi_rank}{country rank according to corruption perception index}
\item{human_development_index}{lower values reflect lowerdevelopment}
\item{hdi_rank}{country rank according to human development index}
\item{year}{year 2015}
}}
\source{
\itemize{
\item \href{http://hdr.undp.org/en/data}{UNDP}
\item \href{https://www.transparency.org/cpi2015#downloads}{Transparency.org}
\item \href{https://docs.google.com/spreadsheet/pub?key=phAwcNAVuyj0XOoBL_n5tAQ&output=xlsx}{Gapminder population data}
}
}
\usage{
hdi_cpi_2015
}
\description{
A dataframe combining Human Development Index by UNDP and Corruption Perception Index from Transparency International for 121 countries.
}
\keyword{datasets}
