% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_cpi_2015.R
\docType{data}
\name{hdi_cpi_2015}
\alias{hdi_cpi_2015}
\title{hdi_cpi_2015}
\format{The dataframe \code{hdi_cpi_2015} has 121 rows and 10 variables:
\itemize{
\item \strong{country}: factor with 121 levels
\item \strong{country_code}: factor
\item \strong{region}: factor with 6 levels
\item \strong{continent}: factor with 5 levels
\item \strong{population}: country population in 2015
\item \strong{corruption_perception_index}:  measured index with lower values reflecting higher perceived corruption
\item \strong{cpi_rank}:  country rank according to corruption perception index
\item \strong{human_development_index}:  measured index with lower values reflecting lower development
\item \strong{hdi_rank}:  country rank according to human development index
\item \strong{year}:  year 2015
}}
\source{
\itemize{
\item \href{http://hdr.undp.org/en/data}{United Nations Development Program}
\item \href{https://www.transparency.org/cpi2015#downloads}{Transparency.org}
\item \href{https://docs.google.com/spreadsheet/pub?key=phAwcNAVuyj0XOoBL_n5tAQ&output=xlsx}{Gapminder population data}
}
}
\usage{
hdi_cpi_2015
}
\description{
A dataframe combining Human Development Index by the United Nations Development Program(UNDP) and Corruption Perception Index from Transparency International for 121 countries.
}
\keyword{datasets}
